apport: auto collect logs for bug reports
